reviews on detail for now

+++ database/migrations/2024_01_10_101653_create_reviews_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string("review_user", 100);
             $table->foreign("review_user")->references("username")->on("users");
             $table->foreignId("review_item_id")->constrained("items", "item_id");
-            $table->decimal("review_rating", 2, 1);
+            $table->tinyInteger("review_rating");
             $table->string("review_text", 500);
             $table->timestampsTz();
         });
